Version 0.2
In this version I made sure that Numi is running when the action is
started. If not, the action starts Numi in background. It is also now
possible to start the action without input which activates Numi. LICENSE
is added to reflect permission by Numi developer to distribute this
action.